+++ modules/integration/fsx-lustre-on-eks/tests/test_stack.py
@@ -10,9 +10,9 @@ import pytest
 
 @pytest.fixture(scope="function")
 def stack_defaults():
-    os.environ["ADDF_PROJECT_NAME"] = "test-project"
-    os.environ["ADDF_DEPLOYMENT_NAME"] = "test-deployment"
-    os.environ["ADDF_MODULE_NAME"] = "test-module"
+    os.environ["SEEDFARMER_PROJECT_NAME"] = "test-project"
+    os.environ["SEEDFARMER_DEPLOYMENT_NAME"] = "test-deployment"
+    os.environ["SEEDFARMER_MODULE_NAME"] = "test-module"
     os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
     os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
 
@@ -24,12 +24,14 @@ def test_synthesize_stack(stack_defaults):
     import stack_fsx_eks
 
     app = cdk.App()
+    proj_name = "addf"
     dep_name = "test-deployment"
     mod_name = "test-module"
 
     stack_fsx_eks.FSXFileStorageOnEKS(
         scope=app,
-        id=f"addf-{dep_name}-{mod_name}",
+        id=f"{proj_name}-{dep_name}-{mod_name}",
+        project_name=proj_name,
         deployment_name=dep_name,
         module_name=mod_name,
         eks_cluster_name="myekscluster",